Add tests, move most to integration
